requirements.txt
aiofile==3.7.4
aiofiles==0.8.0
aioredis==2.0.1
alembic==1.8.0
anyio==3.6.1
arq @ git+https://github.com/samuelcolvin/arq@4c70063bd62046fb94f0a0b5096bf0a35b62b26a
async-timeout==4.0.2
async-unzip==0.2.1
asyncpg==0.25.0
caio==0.9.6
certifi==2022.6.15
click==8.1.3
Deprecated==1.2.13
gino==1.0.1
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
humanize==4.2.3
idna==3.3
ijson==3.1.4
Mako==1.2.1
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
pyaml==21.10.1
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
rfc3986==1.5.0
sanic==22.6.0
sanic-routing==22.3.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.24
typing_extensions==4.3.0
ujson==5.4.0
uvloop==0.16.0
websockets==10.3
wrapt==1.14.1
